requirements.txt
aiohttp==3.11.10
beautifulsoup4==4.12.3
discord.py==2.4.0
feedparser==6.0.11
lxml==5.3.0
openai==1.12.0
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.67.1
